feat/*: move global styles & ad animations file
