+++ src/App.js
@@ -1,14 +1,16 @@
 import React from 'react';
 import { StoryRow } from './components/StoryRow/index'
-import { GlobalStyles } from './globalStyles';
+import { GlobalStyles } from './styles/globalStyles';
 import { PhotoCardList } from './components/PhotoCardList/index';
+import { Logo } from './components/Logo/index'
 
 
 export const App = () => (
     <div>
-        <GlobalStyles/>
-        <StoryRow/>
-        <PhotoCardList/>
+        <GlobalStyles />
+        <Logo />
+        <StoryRow />
+        <PhotoCardList />
     </div>
     
 );
