+++ src/styles/animation.js
@@ -1,0 +1,16 @@
+import { keyframes, css } from 'styled-components'
+
+const fadeInKeyframes = keyframes`
+  from {
+    filter: blur(5px);
+    opacity: 0;
+  }
+
+  to {
+    filter: blur(0);
+    opacity: 1;
+  }
+`
+
+export const fadeIn = ({ time = '1s', type = 'ease' } = {}) =>
+  css`animation: ${time} ${fadeInKeyframes} ${type};`
